pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "Just a sample blog"
authors = ["Sosshi <symonmwenex@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
Pillow = "^8.4.0"
psycopg2-binary = "^2.9.1"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
